+++ src/ts/markdown/mermaidRender.ts
@@ -1,10 +1,11 @@
 import {Constants} from "../constants";
 import {addScript} from "../util/addScript";
 import {mermaidRenderAdapter} from "./adapterRender";
+import {genUUID} from "../util/function";
 
 declare const mermaid: {
     initialize(options: any): void,
-    init(options: any, element: Element): void,
+    render(id: string, text: string): { svg: string }
 };
 
 export const mermaidRender = (element: HTMLElement, cdn = Constants.CDN, theme: string) => {
@@ -26,7 +27,8 @@ export const mermaidRender = (element: HTMLElement, cdn = Constants.CDN, theme: 
                 useMaxWidth: true,
                 diagramMarginX: 8,
                 diagramMarginY: 8,
-                boxMargin: 8
+                boxMargin: 8,
+                showSequenceNumbers: true // Mermaid 时序图增加序号 https://github.com/siyuan-note/siyuan/pull/6992 https://mermaid.js.org/syntax/sequenceDiagram.html#sequencenumbers
             },
             gantt: {
                 leftPadding: 75,
@@ -37,12 +39,21 @@ export const mermaidRender = (element: HTMLElement, cdn = Constants.CDN, theme: 
             config.theme = "dark";
         }
         mermaid.initialize(config);
-        mermaidElements.forEach((item) => {
+        mermaidElements.forEach(async (item) => {
             const code = mermaidRenderAdapter.getCode(item);
             if (item.getAttribute("data-processed") === "true" || code.trim() === "") {
                 return;
             }
-            mermaid.init(undefined, item);
+            const id = "mermaid"+genUUID()
+            try {
+                const mermaidData = await mermaid.render(id, item.textContent);
+                item.innerHTML = mermaidData.svg;
+            } catch (e) {
+                const errorElement = document.querySelector("#" + id);
+                item.innerHTML = `${errorElement.outerHTML}<br>
+<div style="text-align: left"><small>${e.message.replace(/\n/, "<br>")}</small></div>`;
+                errorElement.parentElement.remove();
+            }
             item.setAttribute("data-processed", "true");
         });
     });